-Add AnimatedTileEditor
