+++ Editor/Tiles/AnimatedTile.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 002d0dba65a5470db8f0ec135f0288b1
+timeCreated: 1673494987
