+++ Editor/Tiles/AnimatedTile/AnimatedTileEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f828a183ef5d4eb5adf399bfc893d4f4
+timeCreated: 1673494999
